requirements.txt
google-assistant-grpc==0.0.2
grpc-google-cloud-speech-v1beta1==0.14.0
google-auth-oauthlib==0.1.0
slackclient
